chore: path src public to root (#1)
