+++ index.js
@@ -1,23 +1,21 @@
-const express = require('express'); //
+const express = require("express"); //
 const app = express();
-const morgan = require('morgan'); //
+const morgan = require("morgan"); //
 const cors = require("cors");
-const router = require("./routes/index");
-const pool = require('./database/database.js');
+const router = require("./src/routes/index");
+const pool = require("./src/database/database.js");
 
-app.set('port', 4000);
+app.set("port", 4000);
 
 //middlewares
-app.use(morgan('dev')); 
+app.use(morgan("dev"));
 
-
-//routes 
-app.use(express.static(__dirname + '/src/public'))
+//routes
+app.use(express.static(__dirname + "/public"));
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 
-
-//iniciar 
-app.listen(app.get('port'),()=>{  
- console.log('Listening on port ',app.get('port'));
+//iniciar
+app.listen(app.get("port"), () => {
+  console.log("Listening on port ", app.get("port"));
 });
