+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node index.js",
-    "dev": "nodemon src/"
+    "dev": "nodemon ./"
   },
   "keywords": [],
   "author": "FFJ",
