+++ public/index.html
@@ -27,7 +27,7 @@
                 <a class="nav-link active" aria-current="page" href="info.html">informacion</a>
               </li>
               <li class="nav-item">
-                <a class="nav-link" href="registrarse.html">registrarse</a>
+                <a class="nav-link" href="signup.html">registrarse</a>
               </li>
               <li class="nav-item">
                 <a class="nav-link" href="login.html">iniciar sesion</a>
