+++ public/info.html
@@ -10,7 +10,7 @@
 
     <link href="style.css" rel="stylesheet">
     
-    <title>MAPA</title>
+    <title>INFORMACION</title>
   </head>
   
   <body>
@@ -26,7 +26,7 @@
                 <a class="nav-link active" aria-current="page" href="info.html">informacion</a>
               </li>
               <li class="nav-item">
-                <a class="nav-link" href="registrarse.html">registrarse</a>
+                <a class="nav-link" href="signup.html">registrarse</a>
               </li>
               <li class="nav-item">
                 <a class="nav-link" href="login.html">iniciar sesion</a>
